Conflicts:
	Surrogat/Surrogat/Areas/CentralBank/Repository/BillRepository.cs
	Surrogat/Surrogat/Surrogat.csproj

+++ Surrogat/Surrogat/Areas/CentralBank/Services/IOrderIBillService.cs
@@ -4,6 +4,6 @@
 
     public interface IOrderIBillService
     {
-        BillDto OrderBill (int issuerId, double amount);
+        BillDto OrderBill (int issuerId, decimal amount);
     }
 }
